+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.0.6
+ - Updated Token with a `CanAccessPersonalData` boolean.
+ - Moved Authorization code and added Client credentials auth.
+
 ## v0.0.5
  - Added TrackApi
  - Added Token.Make()